requirements.txt
coloredlogs==14.0 # pretty print logging
gitpython==3.1.3 # clone git repos
